refactor: move useHoverTimeout to a separate file
